+++ packages/hash/src/index.d.ts
@@ -1,0 +1,1 @@
+export default function murmur2(str: string): string