+++ packages/react/src/emotion-element.js
@@ -59,6 +59,8 @@ export const createEmotionProps = (
   return newProps
 }
 
+const Noop = () => null
+
 let Emotion = /* #__PURE__ */ withEmotionCache(
   /* <any, any> */ (props, cache, ref) => {
     let cssProp = props.css
@@ -90,9 +92,7 @@ let Emotion = /* #__PURE__ */ withEmotionCache(
     let serialized = serializeStyles(
       registeredStyles,
       undefined,
-      typeof cssProp === 'function' || Array.isArray(cssProp)
-        ? React.useContext(ThemeContext)
-        : undefined
+      React.useContext(ThemeContext)
     )
 
     if (
@@ -125,6 +125,7 @@ let Emotion = /* #__PURE__ */ withEmotionCache(
     newProps.className = className
 
     const ele = React.createElement(type, newProps)
+    let possiblyStyleElement = <Noop />
     if (!isBrowser && rules !== undefined) {
       let serializedNames = serialized.name
       let next = serialized.next
@@ -132,20 +133,23 @@ let Emotion = /* #__PURE__ */ withEmotionCache(
         serializedNames += ' ' + next.name
         next = next.next
       }
-      return (
-        <>
-          <style
-            {...{
-              [`data-emotion`]: `${cache.key} ${serializedNames}`,
-              dangerouslySetInnerHTML: { __html: rules },
-              nonce: cache.sheet.nonce
-            }}
-          />
-          {ele}
-        </>
+      possiblyStyleElement = (
+        <style
+          {...{
+            [`data-emotion`]: `${cache.key} ${serializedNames}`,
+            dangerouslySetInnerHTML: { __html: rules },
+            nonce: cache.sheet.nonce
+          }}
+        />
       )
     }
-    return ele
+    // Need to return the same number of siblings or else `React.useId` will cause hydration mismatches.
+    return (
+      <>
+        {possiblyStyleElement}
+        {ele}
+      </>
+    )
   }
 )