+++ packages/react/src/utils.js
@@ -1,3 +1,3 @@
 export let isBrowser = typeof document !== 'undefined'
 
-export const hasOwnProperty = Object.prototype.hasOwnProperty
+export const hasOwnProperty = {}.hasOwnProperty