+++ packages/sheet/__tests__/index.js
@@ -15,6 +15,11 @@ afterEach(() => {
   jest.clearAllMocks()
 })
 
+beforeEach(() => {
+  safeQuerySelector('head').innerHTML = ''
+  safeQuerySelector('body').innerHTML = ''
+})
+
 describe('StyleSheet', () => {
   it('should be speedy by default in production', () => {
     process.env.NODE_ENV = 'production'
@@ -93,7 +98,6 @@ describe('StyleSheet', () => {
     expect(sheet.tags).toHaveLength(1)
     expect(sheet.tags[0].parentNode).toBe(container)
     sheet.flush()
-    document.body.removeChild(container)
   })
 
   it('should accept prepend option', () => {
@@ -108,7 +112,44 @@ describe('StyleSheet', () => {
     expect(document.documentElement).toMatchSnapshot()
 
     sheet.flush()
-    head.removeChild(otherStyle)
+  })
+
+  it('should accept insertionPoint option', () => {
+    const head = safeQuerySelector('head')
+
+    head.innerHTML = `
+      <style id="first"></style>
+      <style id="last"></style>
+    `
+
+    // the sheet should be inserted between the first and last style nodes
+    const sheet = new StyleSheet({
+      ...defaultOptions,
+      insertionPoint: safeQuerySelector('#first')
+    })
+    sheet.insert(rule)
+    sheet.insert(rule2)
+    expect(document.documentElement).toMatchSnapshot()
+
+    sheet.flush()
+  })
+
+  it('should work if insertionPoint is last element', () => {
+    const head = safeQuerySelector('head')
+    const lastStyle = document.createElement('style')
+    lastStyle.setAttribute('id', 'last')
+    head.appendChild(lastStyle)
+
+    // the sheet should be inserted after the first node
+    const sheet = new StyleSheet({
+      ...defaultOptions,
+      insertionPoint: lastStyle
+    })
+    sheet.insert(rule)
+    sheet.insert(rule2)
+    expect(document.documentElement).toMatchSnapshot()
+
+    sheet.flush()
   })
 
   it('should be able to hydrate styles', () => {
@@ -173,6 +214,17 @@ describe('StyleSheet', () => {
     expect(document.documentElement).toMatchSnapshot()
 
     sheet.flush()
-    head.removeChild(otherStyle)
+  })
+
+  it('should not crash when flushing when styles are already detached', () => {
+    const head = safeQuerySelector('head')
+
+    const sheet = new StyleSheet(defaultOptions)
+
+    sheet.insert(rule)
+
+    head.innerHTML = ''
+
+    expect(() => sheet.flush()).not.toThrowError()
   })
 })