+++ packages/cache/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @emotion/cache
 
+## 11.6.0
+
+### Minor Changes
+
+- [#2521](https://github.com/emotion-js/emotion/pull/2521) [`516fe458`](https://github.com/emotion-js/emotion/commit/516fe458058c9ec8218740472b301e935801ebbc) Thanks [@mnajdova](https://github.com/mnajdova)! - Added `insertionPoint` option to the `createCache`. It can be used to insert rules after the specified element. For example, to use it with the `CacheProvider` from `@emotion/react` you can do this:
+
+  ```jsx
+  const head = document.querySelector('head')
+
+  // <meta name="emotion-insertion-point" content="">
+  const emotionInsertionPoint = document.createElement('meta')
+  emotionInsertionPoint.setAttribute('name', 'emotion-insertion-point')
+  emotionInsertionPoint.setAttribute('content', '')
+
+  head.appendChild(emotionInsertionPoint)
+
+  // the emotion sheets should be inserted right after the meta tag
+  const cache = createCache({
+    key: 'my-app',
+    insertionPoint: emotionInsertionPoint
+  })
+
+  function App() {
+    return (
+      <CacheProvider value={cache}>
+        <Main />
+      </CacheProvider>
+    )
+  }
+  ```
+
+### Patch Changes
+
+- Updated dependencies [[`9e82a991`](https://github.com/emotion-js/emotion/commit/9e82a991624b18c20c46c5974e8a127c94a54711)]:
+  - @emotion/sheet@1.1.0
+
+## 11.5.0
+
+### Patch Changes
+
+- [#2498](https://github.com/emotion-js/emotion/pull/2498) [`e5beae8e`](https://github.com/emotion-js/emotion/commit/e5beae8e320f3d1455e45efecdfeb7d757687a43) Thanks [@Andarist](https://github.com/Andarist)! - Fixed an edge case issue with incorrect rules being generated. When a context selector (`&`) was used not at the beginning of a selector (which is not valid SCSS but is allowed by the Stylis parser that we are using) within a group of selectors containing a pseudoclass then it was not replaced correctly with the current context selector.
+
+* [#2493](https://github.com/emotion-js/emotion/pull/2493) [`9ae4a91a`](https://github.com/emotion-js/emotion/commit/9ae4a91a08a6f7c5ca26a585f1c271a179db4623) Thanks [@Andarist](https://github.com/Andarist)! - Improved the declared types related to Stylis. This should improve the compatibility with the types published in `@types/stylis`.
+
+* Updated dependencies [[`f2eda829`](https://github.com/emotion-js/emotion/commit/f2eda8295429dd1892a06cbc9496321f2a55c10a)]:
+  - @emotion/sheet@1.0.3
+
 ## 11.4.0
 
 ### Patch Changes